refactor: move cancellableFetch to http.util
