+++ packages/back-end/src/events/handlers/webhooks/EventWebHookNotifier.ts
@@ -11,13 +11,13 @@ import { findOrganizationById } from "../../../models/OrganizationModel";
 import { createEventWebHookLog } from "../../../models/EventWebHookLogModel";
 import { logger } from "../../../util/logger";
 import {
-  cancellableFetch,
   EventWebHookErrorResult,
   EventWebHookResult,
   EventWebHookSuccessResult,
   getEventWebHookSignatureForPayload,
   getPayloadForNotificationEvent,
 } from "./event-webhooks-utils";
+import { cancellableFetch } from "../../../util/http.util";
 
 let jobDefined = false;
